# spec/api.yml
openapi: 3.0.3
info:
  title: 診所預約系統 API
  description: |
    診所預約系統的 RESTful API 規格文件。

    每個 API 的 summary 欄位對應 Feature File 中的 step statement。
  version: 1.0.0
  contact:
    name: Clinic Booking System
servers:
  - url: /api
    description: API Base URL
...
paths:
...
  /patient/profile:
    get:
      tags:
        - 病患資料處理
      summary: 病患進入個人資料頁面
      description: |
        **Feature:** 病患資料處理.feature

        **Rules:**
        - 以 LINE User ID + 身分證字號作為唯一識別
        - 回診自動帶入歷史資料
      operationId: getPatientProfile
      parameters:
        - name: line_user_id
          in: query
          required: true
          schema:
            type: string
        - name: national_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 姓名欄位預填 {name}，電話欄位預填 {phone}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientProfileResponse'
        '404':
          description: 無歷史資料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - 病患資料處理
      summary: 病患提交個人資料
      description: |
        **Feature:** 病患資料處理.feature

        **Rules:**
        - 姓名必須為 2-20 字元
        - 電話必須為台灣手機格式
        - 身分證字號必須符合台灣身分證格式
        - 出生年月日不可為未來日期
        - 以 LINE User ID + 身分證字號作為唯一識別
        - 首次預約建立新病患資料
      operationId: submitPatientProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientProfileRequest'
      responses:
        '200':
          description: 個人資料被儲存
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientProfileResponse'
        '201':
          description: 建立新病患記錄
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientProfileResponse'
        '400':
          description: 操作失敗 - 資料驗證錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  # ============================================
  # 病患查看預約.feature
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      type: object
      required:
        - error
        - message
        - details
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    PatientProfileRequest:
      type: object
      required:
        - line_user_id
        - name
        - phone
        - national_id
        - birth_date
      properties:
        line_user_id:
          type: string
        name:
          type: string
          minLength: 2
          maxLength: 20
          description: 姓名必須為 2-20 字元
        phone:
          type: string
          pattern: '^09\d{8}$'
          description: 電話必須為台灣手機格式
        national_id:
          type: string
          pattern: '^[A-Z][12]\d{8}$'
          description: 身分證字號必須符合台灣身分證格式
        birth_date:
          type: string
          format: date
          description: 出生年月日不可為未來日期
    PatientProfileResponse:
      type: object
      properties:
        id:
          type: string
        line_user_id:
          type: string
        name:
          type: string
        phone:
          type: string
        national_id:
          type: string
        birth_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
...